CourseIO now works
Course renamed to terrain
Course is now a set of terrains

Terrain splitted into Terrain and TerrainModel to be able to save only the attributes of the terrain with an object stream.

SubMenu class created to have a uniform pattern of title and return buttons for all submenus.
TerrainCreatorMenu updated.

EditorMenu to hold all terrain and course creating options.